successfully introduced redux
